recursion_print_out_a_number
Suppose we have a positive integer, n, that we wish to print out. Our routine will have the heading print_out(n). Assume that the only
I/O routines available will take a single-digit number and output it to the terminal. We will call this routine print_digit; for example,
print_digit(4) will output a 4 to the terminal.